Streamlit/app.py
# Import required modules
import streamlit as st
from PIL import Image

# Import your deep learning model here
# For example, the model might be stored in a .pkl file and can be loaded using joblib or pickle
# from joblib import load
# model = load('model.pkl')
# If the model is a keras model saved in h5 format:
# from tensorflow.keras.models import load_model
# model = load_model('model.h5')

def predict(image):
    """
    Function to perform prediction using the loaded model.
    The image passed to this function is first preprocessed 
    to match the input requirements of the model. 

    If your model expects a certain size, color scale, or 
    normalized data, you would apply those transformations here.

    For example, if you were using a Convolutional Neural Network (CNN) 
    with Keras, it might look like:

    from tensorflow.keras.preprocessing import image as img_prep
    image = img_prep.img_to_array(image)
    image = np.expand_dims(image, axis=0)
    image = image / 255.0
    prediction = model.predict(image)

    The result of model.predict will often be an array of probabilities 
    corresponding to the classes the model can predict.
    You might need to apply a threshold or take the argmax to get a 
    binary prediction or the class label respectively.
    """
    return 'No Nodule Detected'  # Placeholder for demonstration

def load_image(image_file):
    """
    Function to open the image file uploaded by the user using PIL, 
    which stands for Python Imaging Library.

    If there are any universal transformations that need to be made 
    (such as converting to grayscale or resizing), 
    they would be applied here.
    """
    img = Image.open(image_file)
    return img

def landing_page():
    """
    Landing page of the application where the user is welcomed and 
    given a brief overview of the purpose of the application.
    """
    st.title('Welcome to FastVision.ai')
    st.write("""
    Upload a lung image for nodule detection.
    """)
    # The following code can be uncommented if a button is desired to proceed to the prediction page
    # if st.button('Continue'):
    #     return True
    # return False

def prediction_page():
    """
    Page where the user can upload an image and the prediction is displayed.
    After the image is uploaded, it is loaded, displayed, 
    passed to the model for prediction, and the result is displayed.
    """
    st.title('Lung Nodule Detection')

    # File uploader allows the user to upload an image file
    image_file = st.file_uploader("Upload Image", type=['jpg', 'png', 'jpeg'])

    if image_file is not None:
        # Display the uploaded image
        st.image(image_file, use_column_width=True)
        
        # Load the image and perform the prediction
        image = load_image(image_file)
        prediction = predict(image)

        # Show the prediction result
        st.success('Model prediction: {}'.format(prediction))

# If this python file is the main module, execute the landing page function
# and if a False (indicating that the user has not clicked the 'Continue' button) 
# is not returned, execute the prediction page function
if __name__ == "__main__":
    if not landing_page():
        prediction_page()
