Streamlit/requirements.txt
streamlit
Pillow
opencv-python
